requirements.txt
mistralai>=0.0.7
python-dotenv>=1.0.0
pdflatex>=0.1.3
SQLAlchemy>=2.0.0
schedule>=1.2.0
python-dateutil>=2.8.2
jinja2>=3.1.2
pytest>=7.4.0
requests>=2.31.0
